♻️ Prepare for code-splitting in $lib/metadata
