+++ src/lib/state.svelte.js
@@ -6,7 +6,7 @@ import { SvelteMap, SvelteSet } from 'svelte/reactivity';
 import { MetadataInferOptionsNeural } from '$lib/schemas/metadata.js';
 
 import { tables } from './idb.svelte.js';
-import { getMetadataValue } from './metadata.js';
+import { getMetadataValue } from './metadata/index.js';
 
 /**
  * @import * as DB from './database';
