+++ src/routes/(app)/protocols/[id]/metadata/[metadata]/infos/+page.svelte
@@ -11,7 +11,7 @@
 	import FieldUrl from '$lib/FieldURL.svelte';
 	import { uppercaseFirst } from '$lib/i18n';
 	import IconDatatype from '$lib/IconDatatype.svelte';
-	import { MERGEABLE_METADATA_TYPES } from '$lib/metadata.js';
+	import { MERGEABLE_METADATA_TYPES } from '$lib/metadata';
 	import RadioButtons from '$lib/RadioButtons.svelte';
 	import {
 		METADATA_MERGE_METHODS,
