+++ src/routes/(app)/protocols/[id]/metadata/[metadata]/options/+layout.js
@@ -1,5 +1,5 @@
 import { dependencyURI, list } from '$lib/idb.svelte.js';
-import { metadataOptionsKeyRange } from '$lib/metadata.js';
+import { metadataOptionsKeyRange } from '$lib/metadata/index.js';
 import { namespacedMetadataId } from '$lib/schemas/metadata.js';
 
 export async function load({ params, parent, depends }) {
