+++ src/routes/(app)/protocols/[id]/metadata/[metadata]/options/[option]/+page.js
@@ -1,7 +1,7 @@
 import { error } from '@sveltejs/kit';
 
 import { get, list, tables } from '$lib/idb.svelte.js';
-import { metadataOptionsKeyRange } from '$lib/metadata.js';
+import { metadataOptionsKeyRange } from '$lib/metadata/index.js';
 import { metadataOptionId, namespacedMetadataId } from '$lib/schemas/metadata.js';
 import { entries } from '$lib/utils.js';
